+++ CHANGELOG.md
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 - Project name changed to "restful_budget_api"
+- Unit tests of assets replaced with test for expenses
 
 ### Removed